毎日の体重保存

+++ db/migrate/20211121101817_create_healths.rb
@@ -1,7 +1,7 @@
 class CreateHealths < ActiveRecord::Migration[6.0]
   def change
     create_table :healths do |t|
-      t.integer :every_weight, null: false
+      t.float :every_weight, null: false
       t.timestamps
     end
   end
